message and action login && sign in page

+++ src/components/login/signup.js
@@ -1,5 +1,6 @@
-import React, { Component, PropTypes } from 'react'
-import { post } from '../../actions/Api'
+import React, { Component, PropTypes } from 'react';
+import { Link } from 'react-router';
+import { post } from '../../actions/Api';
 
 class Signup extends Component {
 	
@@ -7,7 +8,11 @@ class Signup extends Component {
     super()
 	    this.state = {
 		     email: '',
-		     password: ''
+		     password: '',
+		     valid_email: true,
+		     valid_password: true,
+		     message: '',
+		     success: false
 		 }
     }
     contextTypes: {
@@ -21,36 +26,59 @@ class Signup extends Component {
     		[e.target.name]: e.target.value
     	});
     }
+    messageError = (e, status) => {
+    	if (status) {
+    		this.setState({ 
+    			message: e,
+    			email: '',
+		     	password: '',
+		     	success: 'success'
+		       });
+    	} else {
+    		this.setState({ message: e, success: 'error' });
+    	}
+    }
     // Регестрируем
 	handleSubmit = (e) => {
 	    e.preventDefault()
-	    post('users', this.state).then((user) => {
-	    	return user
-	    }).then((user) => {
-	    	if (user.code === 400) {
-	    		// todo проверка на логин
-	    		return Promise.reject();
-	    	}
-	    	console.log(user);
-	    }).catch((user) => {
-	    	console.log('ошибка' + user)
-	    });
-		window.localStorage.clear();
+	    if (this.state.email.length && this.state.password.length) {
+		    post('users', this.state).then((user) => {
+		    	return user
+		    }).then((user) => {
+		    	if (user.code === 400 || user.code === 403) {
+		    		// todo проверка на логин
+		    		return Promise.reject();
+		    	}
+		    	this.messageError('Пользователь успешно зарегистрирован', true)
+		    }).catch((user) => {
+		    	this.messageError('Такой пользователь уже есть')
+		    });
+			window.localStorage.clear();
+  		} else {
+  			this.setState({
+				valid_email: this.state.email.length ? true : false,
+				valid_password: this.state.password.length ? true : false,
+				message: 'Необходимо заполнить все поля'
+			});
+  		}
   	}
 	render() {
 		return (
 		<div className='o_signup'>
 			<div className='o_container'>
 				<h1>Зарегистрироваться</h1>
-				<form className='o_form o_login__form' onSubmit={this.handleSubmit}>
+				<div className="o_login__form">
+				{this.state.message ? <div className={"o_login__" + this.state.success}>{this.state.message} {this.state.success === 'success' ? <Link to='/panel'>Войти &rarr;</Link> : null }</div> : null }
+				<form className='o_form' onSubmit={this.handleSubmit}>
 	        	  	<div className='o_login__input'>
 	        	  		<input onChange={this.onInputChange} name="email" ref="email" type="email" placeholder='email'/>
 	        	  	</div>
 	        	  	<div className='o_login__input'>
 	        	  		<input onChange={this.onInputChange} name="password" ref="password" type='password' placeholder='password'/>
 	        	  	</div>
-	        	  	<button className='o_login__button' disabled={this.state.email ? false : true} type='submit'>Войти</button>
+	        	  	<button className='o_login__button' disabled={this.state.email ? false : true} type='submit'>Зарегестрироваться</button>
 	        	</form>
+	        	</div>
         	</div>
         </div>
 		);
